Move launch to async in viewmodel
