+++ src/jvmMain/kotlin/io/github/fourlastor/App.kt
@@ -10,7 +10,6 @@ import io.github.fourlastor.ui.JarPicker
 import io.github.fourlastor.ui.Launcher
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
-import java.io.File
 
 @Composable
 @Preview
@@ -28,7 +27,7 @@ fun App(appComponent: AppComponent, getPokeWildsLocation: () -> Pair<String, Str
                         viewModel.angleGles20(it)
                     },
                     runPokeWilds = {
-                        runPokeWilds(it)
+                        viewModel.runPokeWilds(it)
                     }
                 )
 
@@ -36,22 +35,6 @@ fun App(appComponent: AppComponent, getPokeWildsLocation: () -> Pair<String, Str
             }
         }
     }
-}
-
-private fun runPokeWilds(state: SettingsState.Loaded) {
-    val runArgs = mutableListOf("java", "-jar", state.jar).apply {
-        if (state.angleGles20) {
-            add("angle_gles20")
-        }
-        if (state.devMode) {
-            add("dev")
-        }
-    }.toTypedArray()
-    Runtime.getRuntime().exec(
-        runArgs,
-        null,
-        File(state.dir)
-    )
 }
 
 class AppComponent : KoinComponent {
