+++ src/jvmMain/kotlin/io/github/fourlastor/state/ViewModel.kt
@@ -1,5 +1,6 @@
 package io.github.fourlastor.state
 
+import io.github.fourlastor.settings.SettingsState
 import kotlinx.coroutines.flow.StateFlow
 
 interface ViewModel<State> {
@@ -7,6 +8,8 @@ interface ViewModel<State> {
     fun start()
 
     fun stop()
+
+    fun runPokeWilds(state: SettingsState.Loaded)
 
     val state: StateFlow<State>
 }
